% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{candlesticks}
\alias{candlesticks}
\title{Candlestick chart}
\usage{
candlesticks(
  symbol = "BTC",
  currency = "USD",
  start = Sys.Date() - 29,
  end = Sys.Date()
)
}
\arguments{
\item{symbol}{a character string of the symbol. Default is "BTC" for Bitcoin.}

\item{currency}{a character string for the region of the currency. "USD", "EUR", "GBP"... Default is "USD" for US Dollar.}

\item{start}{a date in "YYYY-MM-DD" format. Default is 30 days before today.}

\item{end}{a date in "YYYY-MM-DD" format. Default is today.}
}
\description{
Plots an interactive graph with the candlestick chart of a single cryptocurrency.
}
\examples{
candlesticks()
candlesticks(symbol="LTC",currency="GBP",start="2019-09-01",end="2020-02-10")

}
\author{
Vincent KV \href{mailto:contact@vincentkv.com}{contact@vincentkv.com}
}
