% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preciousFunctions.R
\name{cointrack}
\alias{cointrack}
\title{Coin tracker graph}
\usage{
cointrack(ts)
}
\arguments{
\item{ts}{a time series.}
}
\description{
Plots an interactive graph with the historical data of your portfolio.
}
\examples{
my_portfolio=portfolio(list(wallet(symbol = "TRX",amount = 20,start = "2021-01-01"),wallet(symbol = "LTC",amount = 25,start = "2021-03-01")))
cointrack(my_portfolio)
}
\author{
Vincent KV \href{mailto:contact@vincentkv.com}{contact@vincentkv.com}
}
