% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preciousFunctions.R
\name{cryptoCourse}
\alias{cryptoCourse}
\title{Cryptocurrency course}
\usage{
cryptoCourse(symbol = "BTC", currency = "USD")
}
\arguments{
\item{symbol}{a character string of the symbol. Default is "BTC" for Bitcoin.}

\item{currency}{a character string for the region of the currency. "USD", "EUR", "GBP"... Default is "USD" for US Dollar.}
}
\value{
if found, a time series of the searched cryptocurrency.
}
\description{
Get the daily data (Open, High, Low, Close) of a particular cryptocurrency from its creation until today.
}
\examples{
my_course=cryptoCourse(symbol="LTC")
my_course=cryptoCourse(currency="EUR")
my_course=cryptoCourse(symbol="TRX",currency="GBP")

}
\author{
Vincent KV \href{mailto:contact@vincentkv.com}{contact@vincentkv.com}
}
