% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptolist.R
\docType{data}
\name{cryptolist}
\alias{cryptolist}
\title{Symbols and full names of the cryptocurrencies currently supported by the package.}
\format{
A data frame with 376 rows and 2 variables:
\describe{
\item{symbol}{Symbols of cryptocurrencies}
\item{full_name}{Full names of cryptocurrencies}

}
}
\source{
\url{https://github.com/VincentKV/Cryptics}
}
\usage{
data(cryptolist)
}
\description{
Symbols and full names of the cryptocurrencies currently supported by the package.
}
\keyword{datasets}
