% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{invest}
\alias{invest}
\title{Return on investment for a single cryptocurrency}
\usage{
invest(
  symbol = "BTC",
  currency = "USD",
  amount,
  start,
  end = as.Date(Sys.Date())
)
}
\arguments{
\item{symbol}{a character string of the symbol. Default is "BTC" for Bitcoin.}

\item{currency}{a character string for the region of the currency. "USD", "EUR", "GBP"... Default is "USD" for US Dollar.}

\item{amount}{a numeric value for the amount of money invested, in the previous currency.}

\item{start}{a date in "YYYY-MM-DD" format when the money is invested.}

\item{end}{a date in "YYYY-MM-DD" format when the money is withdrawed. Default is today.}
}
\value{
a numeric value.
}
\description{
Computes the simple gain or loss given a specific cryptocurrency and a particular amount of money invested in it during a period of time.
}
\examples{
my_gain=invest(symbol="BTC",currency="USD",amount=20,start="2021-01-01",end="2021-04-01")

}
\author{
Vincent KV \href{mailto:contact@vincentkv.com}{contact@vincentkv.com}
}
