% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preciousFunctions.R
\name{normalizedCourse}
\alias{normalizedCourse}
\title{Normalized cryptocurrency course}
\usage{
normalizedCourse(ts)
}
\arguments{
\item{ts}{a time series.}
}
\value{
a normalized time series.
}
\description{
Normalizes a given cryptocurrency time series between 0 and 1. The values taken are the "Opens". 1 is the all-time-high.
}
\examples{
my_normalized_course=normalizedCourse(cryptoCourse(symbol="LTC",currency="EUR"))

}
\author{
Vincent KV \href{mailto:contact@vincentkv.com}{contact@vincentkv.com}
}
