% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{portfolio}
\alias{portfolio}
\title{Portfolio}
\usage{
portfolio(walletlist)
}
\arguments{
\item{walletlist}{a list of wallets.}
}
\description{
Computes the time series of your total earnings and losses given a list of wallets.
}
\examples{
my_portfolio=portfolio(list(wallet(symbol = "TRX",amount = 20,start = "2021-01-01"),wallet(symbol = "BTC",amount = 30,start = "2020-07-01",end = "2021-02-01"),wallet(symbol = "LTC",amount = 25,start = "2021-03-01")))

}
\author{
Vincent KV \href{mailto:contact@vincentkv.com}{contact@vincentkv.com}
}
