% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{symbolCheck}
\alias{symbolCheck}
\title{Symbol check}
\usage{
symbolCheck(symbol)
}
\arguments{
\item{symbol}{a character string of the symbol to check in the list. Default is "BTC" for Bitcoin.}
}
\value{
a list containing the full name of the cryptocurrency matching the symbol or a dataframe of proposals.
}
\description{
Check if a cryptocurrency symbol is known from the package or resembles a symbol known from the package.
}
\examples{
symbolCheck("BTC")
symbolCheck("BT")
symbolCheck("XBTCZ")
symbolCheck("XYZ")
}
\author{
Vincent KV \href{mailto:contact@vincentkv.com}{contact@vincentkv.com}
}
