% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preciousFunctions.R
\name{viewAllCryptos}
\alias{viewAllCryptos}
\title{View all cryptocurrencies}
\usage{
viewAllCryptos()
}
\description{
View in a data frame all the cryptocurrencies available in this package.
}
\examples{
viewAllCryptos()

}
\author{
Vincent KV \href{mailto:contact@vincentkv.com}{contact@vincentkv.com}
}
