% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preciousFunctions.R
\name{wallet}
\alias{wallet}
\title{Wallet for a single cryptocurrency}
\usage{
wallet(
  symbol = "BTC",
  currency = "USD",
  amount,
  start,
  end = as.Date(Sys.Date())
)
}
\arguments{
\item{symbol}{a character string of the symbol. Default is "BTC" for Bitcoin.}

\item{currency}{a character string for the region of the currency. "USD", "EUR", "GBP"... Default is "USD" for US Dollar.}

\item{amount}{a numeric value for the amount of money invested, in the previous currency.}

\item{start}{a date in "YYYY-MM-DD" format when the money is invested.}

\item{end}{a date in "YYYY-MM-DD" format when the money is withdrawed. Default is today.}
}
\value{
a time series.
}
\description{
Computes the time series of earnings and losses given a specific cryptocurrency and a particular amount of money invested in it during a period of time.
}
\examples{
my_wallet=wallet(symbol="BTC",currency="USD",amount=20,start="2021-01-01",end="2021-04-01")

}
\author{
Vincent KV \href{mailto:contact@vincentkv.com}{contact@vincentkv.com}
}
