% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preciousFunctions.R
\name{candlesticks}
\alias{candlesticks}
\title{Candlestick chart}
\usage{
candlesticks(
  symbol = "BTC",
  currency = "USD",
  start = Sys.Date() - 30,
  end = Sys.Date(),
  option = "original"
)
}
\arguments{
\item{symbol}{a character string of the symbol. Default is "BTC" for Bitcoin.}

\item{currency}{a character string for the region of the currency. "USD", "EUR", "GBP"... Default is "USD" for US Dollar.}

\item{start}{a date in "YYYY-MM-DD" format. Default is 30 days before today.}

\item{end}{a date in "YYYY-MM-DD" format. Default is today.}

\item{option}{a string : default is "original" and does not alter the data. Change to "norm" to normalize the graph between 0 and 1, 1 is the all-time-high. Change to "log" to obtain the logarithm of the graph.}
}
\description{
Plots an interactive graph with the candlestick chart of a single cryptocurrency.
}
\examples{
candlesticks()
candlesticks(symbol="LTC",currency="GBP",start="2019-09-01",end="2020-02-10")

}
\author{
Vincent KV \href{mailto:contact@vincentkv.com}{contact@vincentkv.com}
}
