% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preciousFunctions.R
\name{historic}
\alias{historic}
\title{Historical data graph}
\usage{
historic(symbol = "BTC", currency = "USD")
}
\arguments{
\item{symbol}{a character string of the symbol. Default is "BTC" for Bitcoin.}

\item{currency}{a character string for the region of the currency. "USD", "EUR", "GBP"... Default is "USD" for US Dollar.}
}
\description{
Plots an interactive graph with the historical data of a single cryptocurrency.
}
\examples{
historic()
historic(symbol="LTC",currency="GBP")

}
\author{
Vincent KV \href{mailto:contact@vincentkv.com}{contact@vincentkv.com}
}
